backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
openai>=1.35.0
python-dotenv==1.0.0
pypdf==3.17.1
python-multipart==0.0.6
docx2txt==0.8
pymupdf==1.23.6
ragas==0.0.22
datasets==2.16.1
llama-index==0.9.48
cryptography
aiosqlite